Move maxinfo_propagator
