+++ csrc/transform_replay.cpp
@@ -17,9 +17,9 @@
 #include <ir/iostream.h>
 #include <ir/utils.h>
 #include <logical_domain_map.h>
-#include <maxinfo_propagator.h>
 #include <ops/arith.h>
 #include <options.h>
+#include <scheduler/tools/maxinfo_propagator.h>
 #include <transform_iter.h>
 
 #include <deque>
